requirements.txt
Django==5.2.5
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9